P4ADEV-1545 moved packages
