+++ src/main/java/it/gov/pagopa/payhub/activities/activity/paymentsreporting/ReportingFlowIngestionActivityImpl.java
@@ -1,8 +1,8 @@
 package it.gov.pagopa.payhub.activities.activity.paymentsreporting;
 
-import it.gov.pagopa.payhub.activities.activity.paymentsreporting.service.IngestionFileHandlerService;
-import it.gov.pagopa.payhub.activities.activity.paymentsreporting.service.IngestionFileValidatorService;
-import it.gov.pagopa.payhub.activities.activity.paymentsreporting.service.IngestionFlowRetrieverService;
+import it.gov.pagopa.payhub.activities.service.IngestionFileHandlerService;
+import it.gov.pagopa.payhub.activities.service.IngestionFileValidatorService;
+import it.gov.pagopa.payhub.activities.service.IngestionFlowRetrieverService;
 import it.gov.pagopa.payhub.activities.dto.reportingflow.ReportingFlowIngestionActivityResult;
 import it.gov.pagopa.payhub.activities.dto.reportingflow.IngestionFlowDTO;
 import lombok.extern.slf4j.Slf4j;
